WebフレームワークEchoでwebsocket melodyを動かす
